requirements.txt
filelock==3.20.0
fsspec==2025.10.0
h5py==3.15.1
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
torchmetrics==0.10.3
tqdm==4.67.1
triton==3.5.0
typing_extensions==4.15.0
tzdata==2025.2